requirements.txt
absl-py==2.1.0
addict==2.4.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.1
Flask-Uploads==0.2.1
flatbuffers==23.5.26
fonttools==4.47.2
fqdn==1.5.1
fsspec==2023.12.2
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
hub-sdk==0.0.3
humanfriendly==10.0
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jstyleson==0.0.2
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.11
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.9
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
ncnn==1.0.20240102
nest-asyncio==1.6.0
networkx==3.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.3
oauthlib==3.2.2
onnx==1.15.0
onnx-graphsurgeon==0.3.27
onnx2tf==1.17.5
onnxruntime==1.16.3
onnxsim==0.4.35
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openvino==2023.2.0
openvino-dev==2023.2.0
openvino-telemetry==2023.2.1
opt-einsum==3.3.0
overrides==7.6.0
packaging==23.2
paddlepaddle==2.6.0
pafy==0.5.5
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
portalocker==2.8.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.2
psutil==5.9.7
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.13.1
Send2Trash==1.8.2
six==1.16.0
sng4onnx==1.0.1
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.0
texttable==1.7.0
tflite-support==0.1.0a1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tinycss2==1.2.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.1.2
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.14.1
x2paddle==1.4.1
youtube-dl==2020.12.2
